test: extract shared BigQuery fakes into tests/bigquery_fakes.py
Both test_store.py and test_prediction_log.py duplicated identical fake
BigQuery SDK classes (LoadJobConfig, TimePartitioning, ScalarQueryParameter,
QueryJobConfig, CompletedJob). Extract into a shared module.

Net: −17 lines, single source of truth for BQ test doubles.
